Integrate and factor out common code from modified-lar. Elimate artificial module hierarchy
