+++ src/App.jsx
@@ -3,12 +3,9 @@ import { Switch, Route } from 'react-router-dom'
 import AppHeader from './common/AppHeader.jsx'
 import Footer from './common/Footer.jsx'
 import Home from './Home.jsx'
-import ModifiedLar from './reports/modified-lar/index.jsx'
-import Disclosure from './reports/disclosure/index.jsx'
-import Aggregate from './reports/aggregate/index.jsx'
-import MsaMds from './reports/MsaMds.jsx'
-import Reports from './reports/Reports.jsx'
-import Report from './reports/Report.jsx'
+import ModifiedLar from './reports/ModifiedLar.jsx'
+import Disclosure from './reports/Disclosure.jsx'
+import Aggregate from './reports/Aggregate.jsx'
 import NotFound from './common/NotFound.jsx'
 
 const App = () => {
@@ -20,28 +17,13 @@ const App = () => {
         <Route path="/modified-lar/*" component={NotFound} />
         <Route path="/modified-lar" component={ModifiedLar} />
         <Route
-          path="/disclosure-reports/institution/:institutionId/msa-md/:msaMdId/report/:reportId"
-          component={Report}
+          path="/disclosure-reports/:institutionId?/:msaMdId?/:reportId?"
+          component={Disclosure}
         />
         <Route
-          path="/disclosure-reports/institution/:institutionId/msa-md/:msaMdId"
-          component={Reports}
+          path="/aggregate-reports/:stateId?/:msaMdId?/:reportId?"
+          component={Aggregate}
         />
-        <Route
-          path="/disclosure-reports/institution/:institutionId"
-          component={MsaMds}
-        />
-        <Route path="/disclosure-reports" component={Disclosure} />
-        <Route
-          path="/aggregate-reports/state/:stateId/msa-md/:msaMdId/report/:reportId"
-          component={Report}
-        />
-        <Route
-          path="/aggregate-reports/state/:stateId/msa-md/:msaMdId"
-          component={Reports}
-        />
-        <Route path="/aggregate-reports/state/:stateId" component={MsaMds} />
-        <Route path="/aggregate-reports" component={Aggregate} />
         <Route component={NotFound} />
       </Switch>
       <Footer />
