+++ src/app.scss
@@ -14,7 +14,6 @@
 @import 'common/Header.scss';
 @import 'common/Footer.scss';
 @import 'common/LoadingIcon.scss';
-@import 'reports/modified-lar/Modified-LAR.scss';
-@import 'reports/modified-lar/Results.scss';
-@import 'reports/disclosure/Disclosure.scss';
+@import 'reports/Results.scss';
+@import 'reports/SearchList.scss';
 @import 'reports/tables/tables.scss';
