+++ src/reports/SearchList.scss
@@ -1,4 +1,4 @@
-.disclosure {
+.SearchList {
   .usa-input-error {
     padding-top: 0;
     padding-bottom: 0;
