--- src/reports/disclosure/NoMatch.jsx
@@ -1,5 +1,0 @@
-import React from 'react'
-
-const NoMatch = ({ match }) => <h1>Not found</h1>
-
-export default NoMatch
