--- src/reports/modified-lar/Modified-LAR.scss
@@ -1,9 +1,0 @@
-.modified-lar {
-  .usa-input-error {
-    padding-top: 0;
-    padding-bottom: 0;
-  }
-  .usa-alert {
-    max-width: 46rem;
-  }
-}
